Clean old static structs from dynamic_link Comment their calls in Xcos

+++ scilab/modules/dynamic_link/sci_gateway/cpp/sci_ulink.cpp
@@ -75,7 +75,7 @@ void unLinkAll()
 /*--------------------------------------------------------------------------*/
 void unLink(int _iLib)
 {
-    unsigned long iLib = ConfigVariable::getDynamicLibrary(_iLib)->hLib;
+    unsigned long long iLib = ConfigVariable::getDynamicLibrary(_iLib)->hLib;
     ConfigVariable::removeDynamicLibrary(_iLib);
     Sci_dlclose(iLib);
 }
